add new tests
tst25d allows to test automatically inputs with up to 25 significant digits.

It turned out that n2_small_strtod produces very low number of rounding
errors only as long as input contains 16 or less significant digints.
With longer inputs the percentege of rounding grows however for 17 to 19 digits
the maximal error remains small. But for inputs with 20 or more digits the
maximal error is no lomger small. It reaches -0.504 ULP.
Which is understandable, because my input handler mostly ignores all but 19
significant digits of the input.