n2_small_strtod - another tweak for smaller size
n2_small_strtod - remove temporary variable succ.
Use endptrval as a replacement - update with current value of parse
ponter p on each detected decimal digit, compare it to str as creterion
of parsing succes. Small additional logic is required to handle a case
of valid input with decimal point as a last digit, but the size of this
logic (on nios2) is smaller than savings (not so on x64,
but size on n2_small_strtod does not matter).