n2_small_strtod - more improvements
Get rid off comparison with non-constant maxVal.
Define maxVal constant to have 32 LS bits == 0, to simplify code,
generated for comparison.
Restructure code to remove remaining gotos.

The size is back to where it was before previous round of changes and even
a little lower, but the code is now easier to follow.
Easier, not easy.