w3_small_sttod.c - apply the same change
w3_small_sttod.c - apply the same change that was applied in the
previous commit to n2_small_sttod.c, trading speed in uncommon cases (deep underflow or overflow)
for improved code size.
In no-mulx case the size of w3 variant is already on par with n2 variant.
In mulx case w3 variant is 68-76 bytes bigger.