+big/gen_test1m/cpp - variant of gen_test1 with command line parameters similar to gen_test3
The purpuse of gen_test1m is to add control of range of generated values (-fmin/-fmax, same as in gen_test3).
Unfortunately, it was impossible to add new options to gen_test1 in both convinient and backward-compatible manner. So, I cloned gen_test1 as gen_test1m and did it in non-compatible way.